moved useExistingWorkflowNames under queries/template.ts
