+++ libs/designer/src/lib/core/queries/template.ts
@@ -1,0 +1,8 @@
+import { TemplateService } from '@microsoft/logic-apps-shared';
+import { useQuery } from '@tanstack/react-query';
+
+export const useExistingWorkflowNames = () => {
+  return useQuery(['getExistingWorkflowNames'], async () => {
+    return await TemplateService()?.getExistingWorkflowNames();
+  });
+};
